Make Pay trait from salaries pallet more generic
