+++ frame/support/src/traits/tokens.rs
@@ -27,7 +27,9 @@ pub mod nonfungible_v2;
 pub mod nonfungibles;
 pub mod nonfungibles_v2;
 pub use imbalance::Imbalance;
+pub mod pay;
 pub use misc::{
 	AssetId, Balance, BalanceConversion, BalanceStatus, ConvertRank, DepositConsequence,
 	ExistenceRequirement, GetSalary, Locker, WithdrawConsequence, WithdrawReasons,
 };
+pub use pay::{Pay, PayFromAccount, PaymentStatus};
